Refactor RegisterPropertiesInspection
