+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/inspections/registration/property/quickfix/RegisterPropertyMutabilityQuickFix.kt
@@ -1,37 +1,16 @@
-package godot.intellij.plugin.inspections
+package godot.intellij.plugin.inspections.registration.property.quickfix
 
 import com.intellij.codeInspection.LocalQuickFix
 import com.intellij.codeInspection.ProblemDescriptor
-import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.notification.NotificationDisplayType
 import com.intellij.notification.NotificationGroup
 import com.intellij.notification.NotificationType
 import com.intellij.notification.Notifications
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.NotNullLazyValue
-import com.intellij.psi.PsiElementVisitor
-import org.jetbrains.kotlin.idea.inspections.AbstractKotlinInspection
-import org.jetbrains.kotlin.idea.util.findAnnotation
-import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.KtPsiFactory
-import org.jetbrains.kotlin.psi.propertyVisitor
 
-class RegisterPropertiesInspection : AbstractKotlinInspection() {
-    private val quickFix = RegisterPropertyMutabilityQuickFix()
-
-    override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
-        return propertyVisitor { ktProperty ->
-            if (ktProperty.findAnnotation(FqName("godot.annotation.RegisterProperty")) != null) {
-                if (!ktProperty.isVar) {
-                    holder.registerProblem(ktProperty.valOrVarKeyword, "Registered properties have to be mutable", quickFix)
-                }
-            }
-        }
-    }
-
-}
-
-private class RegisterPropertyMutabilityQuickFix : LocalQuickFix {
+class RegisterPropertyMutabilityQuickFix : LocalQuickFix {
     override fun getFamilyName(): String {
         return "Make property mutable"
     }
@@ -44,8 +23,8 @@ private class RegisterPropertyMutabilityQuickFix : LocalQuickFix {
             Notifications.Bus.notify(
                 NOTIFICATION_GROUP.value
                     .createNotification(
-                        "Register Property Quick Fix",
-                        "Could not change the mutability of property. Change it manually",
+                        "@RegisterProperty Quick Fix",
+                        "Could not change the mutability of property ${descriptor.psiElement.nextSibling.text}. Change it manually",
                         NotificationType.ERROR,
                         null
                     )
