+++ kt/plugins/godot-intellij-plugin/src/main/resources/META-INF/plugin.xml
@@ -12,7 +12,7 @@
         <projectService serviceImplementation="godot.intellij.plugin.services.MyProjectService" />
         <localInspection enabledByDefault="true"
             level="ERROR"
-            implementationClass="godot.intellij.plugin.inspections.RegisterPropertiesInspection"
+            implementationClass="godot.intellij.plugin.inspections.registration.property.inspector.RegisterPropertiesInspection"
             displayName="Register property inspector" groupKey="registerPropertyInspector" />
     </extensions>
